+++ crates/api_models/src/process_tracker/revenue_recovery.rs
@@ -2,17 +2,19 @@ use crate::enums;
 use common_utils::id_type;
 use serde::{Deserialize, Serialize};
 use time::PrimitiveDateTime;
-#[derive(Debug, Clone, Serialize, Deserialize)]
+use utoipa::ToSchema;
+#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
 pub struct RevenueRecoveryResponse {
     pub id: String,
     pub name: Option<String>,
     pub schedule_time_for_payment: Option<PrimitiveDateTime>,
     pub schedule_time_for_psync: Option<PrimitiveDateTime>,
+    #[schema(value_type = ProcessTrackerStatus, example = "finish")]
     pub status: enums::ProcessTrackerStatus,
     pub business_status: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
 pub struct RevenueRecoveryId {
     pub revenue_recovery_id: id_type::GlobalPaymentId,
 }
