+++ crates/common_utils/src/id_type/global_id/payment.rs
@@ -1,3 +1,4 @@
+use common_enums::enums;
 use error_stack::ResultExt;
 
 use crate::{errors, generate_id_with_default_len, generate_time_ordered_id_without_prefix, types};
@@ -30,6 +31,15 @@ impl GlobalPaymentId {
     /// Generate a new ClientId from self
     pub fn generate_client_secret(&self) -> types::ClientSecret {
         types::ClientSecret::new(self.clone(), generate_time_ordered_id_without_prefix())
+    }
+
+    /// Generate the id for revenue recovery Execute PT workflow
+    pub fn get_execute_revenue_recovery_id(
+        &self,
+        task: &str,
+        runner: enums::ProcessTrackerRunner,
+    ) -> String {
+        format!("{task}_{runner}_{}", self.get_string_repr())
     }
 }
 
@@ -66,6 +76,15 @@ impl GlobalAttemptId {
     /// Get string representation of the id
     pub fn get_string_repr(&self) -> &str {
         self.0.get_string_repr()
+    }
+
+    /// Generate the id for Revenue Recovery Psync PT workflow
+    pub fn get_psync_revenue_recovery_id(
+        &self,
+        task: &str,
+        runner: enums::ProcessTrackerRunner,
+    ) -> String {
+        format!("{task}_{runner}_{}", self.get_string_repr())
     }
 }
