+++ crates/router/src/services/authorization/permissions.rs
@@ -95,7 +95,7 @@ generate_permissions! {
             scopes: [Read, Write],
             entities: [Merchant]
         },
-        ProcessTracker: {
+        RevenueRecovery: {
             scopes: [Read],
             entities: [Merchant]
         }
@@ -113,7 +113,7 @@ pub fn get_resource_name(resource: Resource, entity_type: EntityType) -> &'stati
         (Resource::ApiKey, _) => "Api Keys",
         (Resource::Connector, _) => "Payment Processors, Payout Processors, Fraud & Risk Managers",
         (Resource::Routing, _) => "Routing",
-        (Resource::ProcessTracker, _) => "Process Tracker",
+        (Resource::RevenueRecovery, _) => "Revenue Recovery",
         (Resource::ThreeDsDecisionManager, _) => "3DS Decision Manager",
         (Resource::SurchargeDecisionManager, _) => "Surcharge Decision Manager",
         (Resource::Analytics, _) => "Analytics",
